Repot filter stack impl into macro
